Updated Packetbeat DNS tunnel blog post for 5.x
I updated the example to stay in sync with the blog post. The main change was that packetbeat provides the `dns.question.etld_plus_one` field so we no longer need a scripted field in the aggregation and an analyzed version of `dns.question.name` for whitelisting.

Thanks gingerwizard for changing the scripts from groovy to painless.